isolated the AtomDistribution utility
