+++ regtest/tools/rt-make-AtomicDistribution/Makefile
@@ -1,0 +1,1 @@
+include ../../scripts/test.make
