+++ regtest/tools/rt-make-AtomicDistribution/config
@@ -1,0 +1,1 @@
+type=make
